+++ saas_portal/doc/changelog.rst
@@ -2,6 +2,7 @@
 -------
 
 - **Improvement:** using async counters, that were introduced in https://github.com/odoo/odoo/commit/73b4e9b1b671690a9d235459a6127da3603083a3
+- **New:** seperating temporary builds from main ones
 
 `1.0.1`
 -------